+++ src/routes/modules/hris.route.ts
@@ -4,25 +4,14 @@ import {employeesCsvToJsonArray} from "../../functions/modules/hris.function";
 const hris = express.Router();
 
 hris.post("/import-employees", uploadCsv.single("csv"), async (req, res) => {
-	const expectedHeader = [
-		"lastName",
-		"firstName",
-		"middleName",
-		"suffix",
-		"phoneNumber",
-		"nickname",
-		"email",
-		"birthday",
-		"bloodType",
-		"driverLicense",
-		"taxId",
-	];
+	const currentUser = req.authorization;
 	const csvBuffer = req.file?.buffer.toString("utf8");
+
 	if (!csvBuffer) {
 		res.status(400).json({error: "please provide a csv"});
 	} else {
 		try {
-			const result = await employeesCsvToJsonArray({expectedHeader, csvBuffer});
+			const result = await employeesCsvToJsonArray(csvBuffer, currentUser.companyUuid);
 			res.status(result.status).json(result.data);
 		} catch (error: any) {
 			res.status(500).json({error: error.message});
